:file_folder: refactor: Create package `test/fixtures`
